Change outputs to flow outputs in Overview.vue
The issue was underneath the Overview.vue/executions file.

Since it is a .vue file type, all what was needed to be changed was the <h5></5>.

Originally, it was <h5>{{ $t("outputs") }}</h5>, but it has then been renamed to <h5>{{ $t("flow outputs") }}</h5> instead. This should resolve the issue.